requirements.txt
Django==1.5.5
South==0.8.2
djangorestframework==2.3.8
wsgiref==0.1.2
